+++ example/lib/counter/counter_cube_widget.dart
@@ -2,22 +2,42 @@ import 'package:cubes/cubes.dart';
 import 'package:examplecube/counter/counter_cube.dart';
 import 'package:flutter/material.dart';
 
-class CounterCubeWidget extends CubeWidget<CounterCube> {
+enum CounterCubeKeyAnimation { scaleController, scaleAnimation }
+
+class CounterCubeWidget extends CubeWidgetAnimation<CounterCube> {
+  @override
+  void initState(BuildContext context, CounterCube cube, TickerProvider vsync) {
+    animationControllers[CounterCubeKeyAnimation.scaleController] =
+        AnimationController(
+      vsync: vsync,
+      duration: Duration(seconds: 1),
+    );
+    animations[CounterCubeKeyAnimation.scaleAnimation] = CurvedAnimation(
+      parent: animationControllers[CounterCubeKeyAnimation.scaleController],
+      curve: Curves.bounceOut,
+    );
+    animationControllers[CounterCubeKeyAnimation.scaleController].forward();
+    super.initState(context, cube, vsync);
+  }
+
   @override
   Widget buildView(BuildContext context, CounterCube cube) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Counter with CubeWidget'),
+        title: Text('CubeWidgetAnimation'),
       ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(getString('description_counter')),
-            cube.count.build<int>((value) {
-              return Text(value.toString());
-            }),
-          ],
+      body: ScaleTransition(
+        scale: animations[CounterCubeKeyAnimation.scaleAnimation],
+        child: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Text(getString('description_counter')),
+              cube.count.build<int>((value) {
+                return Text(value.toString());
+              }),
+            ],
+          ),
         ),
       ),
       floatingActionButton: FloatingActionButton(